Modify macros.asm to build on latest ca65
As of May 2nd, 2021, .INCLUDE behaviour inside of macros was modified in
ca65. Now, it requires that the file to be included is known at macro
definition time, and substitutes it then. This breaks the AUDIOPACK
macro, so we work around it by removing it. This is OK since AUDIOPACK
is only used for the pack terminators - we add a new macro just for this
purpose: AUDIOPACKTERMINATOR

The guilty change to ca65: https://github.com/cc65/cc65/commit/b9a3c7888822732a0de92741cfe1a3e1b6bb272f

Further discussion: https://github.com/cc65/cc65/issues/1503

Fixes: https://github.com/Herringway/ebsrc/issues/14